+++ view/adminhtml/web/js/grid/columns/status.js
@@ -8,7 +8,7 @@ define([
         },
 
         getStatus: ({status}) => {
-            return `prontopaga payment ${status}`
+            return `prontopaga payment ${status}`;
         }
     });
 });