+++ view/adminhtml/web/js/grid/columns/validate-payment.js
@@ -1,11 +1,13 @@
 define([
-    "jquery",
-    "Magento_Ui/js/grid/columns/column",
-    "Magento_Ui/js/modal/modal",
-    "uiRegistry",
-    "beautify",
-    "beautifyConfig"
+    'jquery',
+    'Magento_Ui/js/grid/columns/column',
+    'Magento_Ui/js/modal/modal',
+    'uiRegistry',
+    'beautify',
+    'beautifyConfig'
 ], function ($, Column, modal, registry, beautify, beautifyConfig) {
+    'use strict';
+
     return Column.extend({
         defaults: {
             bodyTmpl: 'Improntus_ProntoPaga/grid/cells/validate-payment',
@@ -39,20 +41,21 @@ define([
          * @param {*} transaction_id
          */
         showInfo: function ({transaction_id}) {
-            self = this
+            let self = this;
+
             $.ajax({
                 showLoader: true,
                 url: window.validateUrl,
                 data: { transactionId: transaction_id },
-                type: "POST",
+                type: 'POST',
                 dataType: 'json'
             }).done((response) => {
                 if (response.code === 200) {
-                    response = beautify.js_beautify(response.body.message, beautifyConfig)
-                    registry.get('index = improntus_prontopaga_transaction_listing_data_source').reload({'refresh': true})
+                    response = beautify.js_beautify(response.body, beautifyConfig);
+                    registry.get('index = improntus_prontopaga_transaction_listing_data_source').reload({'refresh': true});
                 }
-                $(`${self.modalElem} ${self.modalContent}`).html(response)
-                $(`${self.modalElem}`).modal('setSubTitle', `uid: ${transaction_id}`)
+                $(`${self.modalElem} ${self.modalContent}`).html(response);
+                $(`${self.modalElem}`).modal('setSubTitle', `uid: ${transaction_id}`);
                 $(self.modalElem).modal('openModal');
             });
         }