+++ view/frontend/web/js/view/payment/prontopaga.js
@@ -9,6 +9,7 @@ define(
     ) {
         'use strict';
         const type = 'prontopaga';
+
         if (window.checkoutConfig.payment[type].active) {
             rendererList.push(
                 {