+++ CHANGELOG.md
@@ -2,3 +2,7 @@ CHANGELOG
 ---------
 ### 1.0.0 (2023-10-04)
 * Initial version
+
+### 1.0.1 (2024-04-12)
+* Updated API request methods and remove 'message'.
+* Updated JS files and apply improvements by eslint.